модуляция кода из задания 4
