+++ js/data.js
@@ -1,0 +1,3 @@
+const str = 'I am a Keks cat.';
+
+export {str};
